refactor(Domain): result 디렉토리에서 분리한다
