+++ src/main/java/lotto/domain/grade/Grade.java
@@ -1,4 +1,4 @@
-package lotto.domain.result;
+package lotto.domain.grade;
 
 public enum Grade {
