+++ src/main/java/lotto/domain/grade/GradeCount.java
@@ -1,4 +1,4 @@
-package lotto.domain.result;
+package lotto.domain.grade;
 
 public class GradeCount {
