+++ src/main/java/lotto/domain/lotto/Lotto.java
@@ -1,4 +1,4 @@
-package lotto.domain.number;
+package lotto.domain.lotto;
 
 import java.util.ArrayList;
 import java.util.List;
