+++ src/test/java/lotto/domain/result/GradeTest.java
@@ -2,6 +2,7 @@ package lotto.domain.result;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import lotto.domain.grade.Grade;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
